backend_fastapi/requirements.txt
# ============================================
# SENTINEL IA v2.0 - Backend FastAPI
# Requirements Clean - Version Testée et Fonctionnelle
# Python 3.14 avec PyTorch 2.9.1 CPU
# ============================================

# ===== ESSENTIELS =====

# Web Framework & ASGI
fastapi==0.121.2
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Authentication & Security
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
pydantic[email]==2.12.4
pydantic-settings==2.12.0
email-validator==2.3.0

# Database
sqlalchemy[asyncio]==2.0.44
aiosqlite==0.21.0

# WebSocket
websockets==15.0.1
python-socketio==5.14.3

# HTTP Client
httpx==0.28.1
aiohttp==3.13.2

# Configuration & Logging
pyyaml==6.0.3
python-dotenv==1.2.1
loguru==0.7.3

# Monitoring
psutil==7.1.3

# ===== AI & VISION =====

# PyTorch CPU (pour Python 3.14)
torch==2.9.1
torchvision==0.24.1

# YOLO & Computer Vision
ultralytics==8.3.228
opencv-python==4.12.0.88
numpy==2.2.6
pillow==12.0.0

# ===== OPTIONNELS =====

# Database - ClickHouse
clickhouse-connect==0.9.1

# Storage - MinIO/S3
aiobotocore==2.25.2
aiofiles>=24.0.0

# Data Processing
polars==1.35.2
